+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## 2.0.0
 
-* **Breaking change** `IsolateChannel` requires a separate import
+**Breaking changes**
+
+* `IsolateChannel` requires a separate import
   `package:stram_channel/isolate_channel.dart`.
   `package:stream_channel/stream_channel.dart` will now not trigger any platform
   concerns due to importing `dart:isolate`.
+* Remove `JsonDocumentTransformer` class. The `jsonDocument` top level is still
+  available.
 
 ## 1.7.0